refactor: Split chatbot into smaller components
