+++ src/App.css
@@ -29,3 +29,16 @@
   font-display: swap;
   src: url("/fonts/InterVariable.woff2") format("woff2");
 }
+
+.chat-scrollbar::-webkit-scrollbar {
+  width: 5px;
+}
+.chat-scrollbar:hover::-webkit-scrollbar-thumb {
+  background-color: #d4d4d8;
+  border-radius: 10px;
+}
+textarea::-webkit-resizer {
+    background-color: transparent;
+    border-width: 0;
+    outline: none;
+  }
